% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Abridge.R
\name{Abridge}
\alias{Abridge}
\title{Abridge population or death data}
\usage{
Abridge(X, value.var = "Deaths")
}
\description{
This function assumes you have columns named Age, Sex, AgeInterval, and Year
}

