% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{addcod}
\alias{addcod}
\title{append a \code{$cod} column if missing}
\usage{
addcod(X)
}
\arguments{
\item{X}{a \code{data.frame}, possibly but not necessarily with column \code{$sex}.}
}
\value{
X with a new column, \code{$cod} appended.
}
\description{
Only handles the case of missing \code{$cod} splitting variable for data of a single year/region. This isn't super robust. If you have many regions or whatever then do it yourself. This function was just written to make \code{ggb()} robust to the case of a user specifying data that don't have any territorial or other subgroups, aside from sex.
}

