% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bh.R
\name{bh1CoverageFromAges}
\alias{bh1CoverageFromAges}
\title{given a set of ages, what is the implied death registration coverage?}
\usage{
bh1CoverageFromAges(codi, agesFit, minA = 15, maxA = 75, eOpen = NULL)
}
\arguments{
\item{codi}{a chunk of data (single sex, year, region, etc) with all columns required by \code{ggb()}}

\item{agesFit}{an integer vector of ages, either returned from \code{ggbgetAgesFit} or user-supplied.}

\item{minA}{the minimum of the age range searched. Default 15}

\item{maxA}{the maximum of the age range searched. Default 75}

\item{eOpen}{optional. A value for remaining life expectancy in the open age group.}
}
\value{
numeric. the estimated level of coverage.
}
\description{
For a single year/sex/region of data (formatted as required by \code{bh1()}), what is the registration coverage implied by a given age range? Called by \code{bh1CoverageFromYear()}. Here, the function simply takes the arithmetic mean of a given age range of \code{$Cx}, as returned by \code{bh1MakeColumns()}.
}

