% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bh.R
\name{bh1MakeColumns}
\alias{bh1MakeColumns}
\title{make the Bennett-Horiuchi quasi lifetable columns required by the estimation method}
\usage{
bh1MakeColumns(codi, minA = 15, maxA = 75, eOmega = NULL)
}
\arguments{
\item{codi}{a chunk of data (single sex, year, region, etc) with all columns required by \code{bh1()}}

\item{minA}{the minimum of the age range searched. Default 15}

\item{maxA}{the maximum of the age range searched. Default 75}

\item{eOmega}{optional. A value for remaining life expectancy in the open age group.}
}
\value{
codi, with many columns added, most importantly \code{$Cx}.
}
\description{
Called by \code{bh1CoverageFromYear()}. This simply modulates some code that would otherwise be repeated. Users probably don't need to call this function directly.
}
\details{

}

