% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bh.R
\name{bh2}
\alias{bh2}
\title{estimate the coverage coefficients}
\usage{
bh2(x, minA = 10, maxA = 75, AgeInt = 5, minAges = 8, sex = "f",
  exact.ages = NULL)
}
\arguments{
\item{x}{}
}
\value{
x
}
\description{
x
}
\details{

}

