% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BrasilFemales}
\alias{BrasilFemales}
\title{Example data for Brasil females by federal states, years 1991 to 2000}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{id}{integer an id number for each state}
\item{pop1}{integer the census population count in 1991}
\item{pop2}{integer the census population count in 2000}
\item{deaths}{numeric average deaths between censuses}
\item{date1}{integer 1991}
\item{date2}{integer 2000}
\item{age}{integer lower age bound for each age group}
\item{sex}{character, 'f'}
}
}
\source{
data downloaded from DATASUS \code{http://www.datasus.gov.br}
}
\usage{
BrasilFemales
}
\description{
A dataset containing 486 rows and 7 variables: Population counts for 1991 and 2000 in abridged ages 0, 1, 5, ... 75, with an open age of 80. Deaths are given as the average death count per age group over the intercensal period. In total there are 53 states in this dataset.
}
\keyword{datasets}
