% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BrasilMales}
\alias{BrasilMales}
\title{Example data for Brasil males by federal states, years 1980 to 1991}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{cod}{integer an id number for each state}
  \item{pop1}{integer the census population count in 1991}
  \item{pop2}{integer the census population count in 2000}
  \item{deaths}{numeric average deaths between censuses}
  \item{year1}{integer 1991}
  \item{year2}{integer 2000}
  \item{age}{integer lower age bound for each age group}
  \item{sex}{character, 'm'}
}}
\source{
data downloaded from DATASUS http://www.datasus.gov.br
}
\usage{
BrasilMales
}
\description{
A dataset containing 486 rows and 7 variables: Population counts for 1980 and 1991 in abridged ages 0, 1, 5, ... 75, with an open age of 80. Deaths are given as the average death count per age group over the ntercensal period. In total there are 53 states in this dataset.
}
\keyword{datasets}
