% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cdmltw.R
\name{cdmltw}
\alias{cdmltw}
\title{Coale-Demeny model lifetable: West}
\usage{
cdmltw(sex = "F")
}
\arguments{
\item{sex}{\code{"F"} or \code{"M"}}
}
\value{
Tons of lifetable output in matrices. Age in columns, lifetable levels in rows.
}
\description{
function from now-deprecated \code{demogR} package. Originally written by Ken Wachter, modified by James Jones, and again by the current maintainer, Tim Riffe. Only minor edits to margin naming in the current version.
}

