% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.R
\name{ddmplot}
\alias{ddmplot}
\title{get a quick overview of the different estimates produced}
\usage{
ddmplot(X, ...)
}
\arguments{
\item{X}{output of \code{ddm()}.}

\item{...}{other arguments passed to \code{plot()}}
}
\value{
called for its graphical device side-effects.
}
\description{
produce a dot plot, where each x position is a unique value of \code{$cod}, and points indicate the GGB, BH1, BH2, and harmonic mean of these. Feed this function the output of \code{ddm()}.
}
