% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{est_birthdays}
\alias{est_birthdays}
\title{helper function to calculate Nx (birthdays)}
\usage{
est_birthdays(pop1, pop2, AgeInt = NULL, nx.method = 2)
}
\arguments{
\item{pop1}{numeric vector of population counts by age from census 1.}

\item{pop2}{numeric vector of population counts by age from census 2.}

\item{AgeInt}{numeric vector of age interval widths}

\item{nx.method}{either 2 or 4. 4 is smoother.}
}
\value{
numeric vector of estiamte birthdays per annum per age
}
\description{
Approximate average birthdays per annum in a reference period and age group between two censuses. We do this by taking geometric means within ages, either from age x in census 1 and age x + 5 in census 2 (\code{nx.method = 2}), or from ages x and x+5 in both censuses. We also divide out the age group width, presumably 5.
}
