% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{ggbAgeTrimSensitivity}
\alias{ggbAgeTrimSensitivity}
\title{calculate and plot ggb results for each possible age trim}
\usage{
ggbAgeTrimSensitivity(
  X,
  minA = 5,
  maxA = 75,
  minAges = 8,
  deaths.summed = FALSE,
  mig.summed = deaths.summed,
  lm.method = "tukey",
  nx.method = 2,
  opt.method = "r2",
  type = "Mxcoverage",
  plot = TRUE
)
}
\arguments{
\item{X}{\code{data.frame} with columns, \code{pop1}, \code{pop2}, \code{deaths}, \code{mig} (optional), \code{date1}, \code{date2}, \code{age}, and \code{id} (if there are more than 1 region/sex/intercensal period).}

\item{minA}{the lowest age to be included in search}

\item{maxA}{the highest age to be included in search (the lower bound thereof)}

\item{minAges}{the minimum number of adjacent ages to be used in estimating}

\item{deaths.summed}{logical. is the deaths column given as the total per age in the intercensal period (\code{TRUE}). By default we assume \code{FALSE}, i.e. that the average annual was given.}

\item{mig.summed}{logical. Is the (optional) net migration column \code{mig} given as the total per age in the intercensal period (\code{TRUE}). By default we assume \code{FALSE}, i.e. that the average annual was given.}

\item{lm.method}{character, one of:\itemize{
\item{\code{"oldschool"}} default sd ratio operation of still unknown origin
\item{\code{"lm"} or \code{"ols"}} for a simple linear model
\item{\code{"tls"}, \code{"orthogonal"}, or \code{"deming"}} for total least squares
\item{\code{"tukey"}, \code{"resistant"}, or "\code{"median"}} for Tukey's resistant line method
}}

\item{nx.method}{either 2 or 4. 4 is smoother.}

\item{opt.method}{what kind of residual do we minimize? choices \code{"RMS"},\code{"logRMS"}, \code{"ORSS"}, \code{"logORSS"} (experimental)}

\item{type}{either \code{"Mxcoverage"} or \code{"resid"}}

\item{plot}{logical. Shall we also render the plot?}
}
\description{
Age trim selection is a big open problem for death distribution methods. Here we can see how the \code{"Mxcoverage"} estiamte varies as a function of lower and upper ages of age trims. Alternatively, one could check other optimal criteria, such as r2, RMS, or ORSS statistics over the same space.
}
\details{
The optimal (using the specified \code{opt.method} criterion) age trim is indicated with a circle. Whether or not \code{plot} is \code{TRUE} this function returns a data object that one may wish to analyze.
}
\examples{
\dontrun{
}
}
