% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggb.R
\name{ggbcoverageFromAges}
\alias{ggbcoverageFromAges}
\title{given a set of ages, what is the implied death registration coverage?}
\usage{
ggbcoverageFromAges(codi, agesfit)
}
\description{
For a single year/sex/region of data (formatted as required by \code{ggb()}), what is the registration coverage implied by a given age range? Called by \code{ggbcoverageFromYear()} and \code{plot.ggb()}.
}

