% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggb.R
\name{ggbgetAgesFit}
\alias{ggbgetAgesFit}
\title{determine the age range that minimizes the mean squared error}
\usage{
ggbgetAgesFit(codi, minA = 15, maxA = 75, minAges = 8)
}
\arguments{
\item{codi}{a chunk of data (single sex, year, region, etc) with all columns required by \code{ggb()}}

\item{minAges}{the minimum number of adjacent ages to be used in estimating}
}
\value{
a vector of ages that minimizes the RMSE
}
\description{
Called by \code{ggbcoverageFromYear()} whenever \code{exact.ages} are not given. This automates what one typically does visually.
}
\details{

}
\seealso{
plot.ggb
}

