% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggb.R
\name{ggbgetResidual}
\alias{ggbgetResidual}
\title{calculate the root means square of the error to help find optimal age range}
\usage{
ggbgetResidual(
  agesi,
  codi,
  lm.method = "oldschool",
  opt.method = "RMSE",
  scale = 1
)
}
\arguments{
\item{agesi}{the vector of ages used for this iteration}

\item{codi}{\code{data.frame} as returned by \code{ggbMakeColumns()}}

\item{lm.method}{character, one of:\itemize{
\item{\code{"oldschool"}} default sd ratio operation of still unknown origin
\item{\code{"lm"} or \code{"ols"}} for a simple linear model
\item{\code{"tls"}, \code{"orthogonal"}, or \code{"deming"}} for total least squares
\item{\code{"tukey"}, \code{"resistant"}, or "\code{"median"}} for Tukey's resistant line method
}}

\item{opt.method}{What should we try to minimize when picking age trims? Current options are \code{"RMSE"}, \code{"ORSS"}, \code{"MAE"}, \code{"MAPE"}, or \code{"r2"}. Default \code{"r2"}.}

\item{scale}{scale factor for the objective}
}
\value{
the RMSE
}
\description{
Called by \code{ggbgetAgesFit()} whenever the user does not want to manually determine the age range used to determine registration coverage. Probably no need to be called by top-level users. If a user would rather determine the optimal age range some other way, then look to\code{ggbcoverageFromYear()}where \code{ggbgetResidual} is called and add another condition or make it call something else.
}
\details{
Given a vector of ages and a line fitting method, we take one of several potential residuals. RMSE is the root of the mean squared error, MAE is the mean absolute error, MAPE is the mean absolute proportional error, ORSS is the standard deviation of the orthogonal residuals, and r2 is the r2 of an OLS fit.
}
