% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggb.R
\name{ggbMakeColumns}
\alias{ggbMakeColumns}
\title{make the growth-adjusted quasi lifetable columns required by GGB method}
\usage{
ggbMakeColumns(codi, minA. = 15, maxA. = 75)
}
\arguments{
\item{codi}{a chunk of data (single sex, year, region, etc) with all columns required by \code{ggb()}}

\item{minA.}{the minimum of the age range searched. Default 15}

\item{maxA.}{the maximum of the age range searched. Default 75}
}
\value{
codi, with many columns added, most importantly \code{$rightterm}, \code{$leftterm}, and \code{$exclude}.
}
\description{
Called by \code{plot.ggb()} and \code{ggbcoverageFromYear()}. This simply modulates some cod ethat would otherwise be repeated. Users probably don't need to call this function directly.
}
\details{

}

