% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh.R
\name{ggbseggetAgesFit}
\alias{ggbseggetAgesFit}
\title{select age range for GGBSEG method automatically}
\usage{
ggbseggetAgesFit(
  codi,
  minA = 15,
  maxA = 75,
  minAges = 8,
  agesFit.ggb = NULL,
  deaths.summed = FALSE
)
}
\arguments{
\item{codi}{a chunk of data (single sex, year, region, etc) with all columns required by \code{ggb()}}

\item{minA}{the lowest age to be included in search}

\item{maxA}{the highest age to be included in search (the lower bound thereof)}

\item{minAges}{the minimum number of adjacent ages to be used in estimating}

\item{agesFit.ggb}{integer vector of age trim from GGB method.}

\item{deaths.summed}{logical. is the deaths column given as the total per age in the intercensal period (\code{TRUE}). By default we assume \code{FALSE}, i.e. that the average annual was given.}
}
\value{
integer vector of ages that minimize the RMSE.
}
\description{
We try all possible age ranges given the constraints provided. Whichever age range has the lowest root mean square residual with respect to the corresponding coverage estimate is returned.
}
