% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{group01}
\alias{group01}
\title{group down standard abridged data in child mortality group}
\usage{
group01(X)
}
\arguments{
\item{X}{standard input as required by \code{ddm()}, \code{ggb()}, \code{bh1()}, or \code{bh2()}}
}
\value{
X, with child ages grouped as necessary (or not)
}
\description{
We want 5-year age groups starting from 0. Standard abridged data has 0i,1,5. So we need to group together 0 and 1. Just for the sake of getting comparable results.
}
