% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{guessDeathsColumn}
\alias{guessDeathsColumn}
\title{Figure out which column is the deaths column}
\usage{
guessDeathsColumn(X)
}
\arguments{
\item{X}{\code{data.frame} that we want to check for a deaths column}
}
\value{
The same \code{data.frame}, returned, with the deaths column renamed as \code{"deaths"}
}
\description{
This function will pick up \code{"death"}, \code{"deaths"}, \code{"Death"}, or \code{"Deaths"} (and maybe some others?) and rename it \code{"deaths"} for easier internal usage.
}
\details{

}

