% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{headerPrep}
\alias{headerPrep}
\title{a utility function to prep the header}
\usage{
headerPrep(X)
}
\arguments{
\item{X}{}
}
\value{
a list of codi chunks (by intercensal period, region, etc), with standardized names, dates, etc.
}
\description{
This is an internal utility function, to save on redundant lines of code. Not so useful for hand-processing.
}
