% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{inUSR}
\alias{inUSR}
\title{does a given pairlist of x and y coordinates fall within the plot region?}
\usage{
inUSR(USR, click)
}
\arguments{
\item{USR, }{as given by \code{par("usr")}}

\item{click}{a pairlist with elements \code{$x} and \code{$y}, as returned by \code{locator(1)}}
}
\value{
logical. \code{TRUE} if in the plot region.
}
\description{
Check to see if a point clicked falls in the plot or outside it. This function is used by \code{ggbChooseAges()}.
}
