% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Moz}
\alias{Moz}
\title{Example data for Mozambique females 1997-2007}
\format{
A data frame with 17 rows and 8 variables:
\describe{
\item{id}{integer a column of 1s}
\item{sex}{character ``f'' for female}
\item{pop1}{integer the census population count in 1997}
\item{pop2}{integer the census population count in 2007}
\item{deaths}{integer average of 1997 and 2007 deaths}
\item{age}{integer lower age bound for each age group}
\item{date1}{integer 1997}
\item{date2}{integer 2007}
}
}
\source{
Data courtesy of Bernardo Queiroz.
}
\usage{
Moz
}
\description{
A dataset containing 17 rows and 8 variables: Population counts for 1997 and 2007 in quinquennial age groups 0, 5, ... 75, with an open age of 80. Deaths are given as the average of the age-specific deaths in 1997 and 2007.
}
\keyword{datasets}
