% File src/library/base/man/load.Rd
\name{project_DDM}
\alias{project_DDM}
\title{Project population using Different Distribution Method (DDM)}
\description{
Project population using Different Distribution Method, The function project_DDM takes input data frame of 16 Age group (00-75+) and two consicutive census population to project future Age Specific Population.
}
\usage{
project_DDM(file)
}
\arguments{
  \item{file}{A "CSV" file which contains the age group and census population information}
}


\author{Sandeep N, Rohith Mohan}


\references{
  Ramnath Takiar (2009) \emph{Cancer incidence rates and the problem of denominators - A new approach in Indian cancer registries}.
}
