% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh.R
\name{segCoverageFromAges}
\alias{segCoverageFromAges}
\title{given a set of ages, what is the implied death registration coverage?}
\usage{
segCoverageFromAges(codi, agesFit)
}
\arguments{
\item{codi}{a chunk of data (single sex, year, region, etc) with all columns required by \code{ggb()}}

\item{agesFit}{an integer vector of ages, either returned from \code{seggetAgesFit} or user-supplied.}
}
\value{
numeric. the estimated level of coverage.
}
\description{
For a single year/sex/region of data (formatted as required by \code{seg()}, \code{ggbseg()}), what is the registration coverage implied by a given age range? Called by \code{segCoverageFromYear()} and  \code{ggbsegCoverageFromYear()}. Here, the function simply takes the arithmetic mean of a given age range of \code{$Cx}, as returned by \code{segMakeColumns()} or \code{ggbsegMakeColumns()}. Not intended for top-level use.
}
