% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh.R
\name{segCoverageFromYear}
\alias{segCoverageFromYear}
\title{estimate death registration coverage for a single year/sex/region using the Bennett-Horiuchi method}
\usage{
segCoverageFromYear(
  codi,
  minA = 15,
  maxA = 75,
  minAges = 8,
  exact.ages = NULL,
  eOpen = NULL,
  nx.method = 2,
  deaths.summed = FALSE,
  mig.summed = deaths.summed,
  delta = FALSE,
  exact.ages.ggb = NULL,
  lm.method = "oldschool",
  opt.method = "r2"
)
}
\arguments{
\item{codi}{\code{data.frame} with columns, \code{$pop1}, \code{$pop2}, \code{$deaths}, \code{$date1}, \code{$date2}, \code{$sex}, \code{$age}, and \code{$id} (to indicate regions, periods, sexes).}

\item{minA}{the lowest age to be included in search}

\item{maxA}{the highest age to be included in search (the lower bound thereof)}

\item{minAges}{the minimum number of adjacent ages to be used in estimating}

\item{exact.ages}{optional. A user-specified vector of exact ages to use for coverage estimation}

\item{eOpen}{optional. A user-specified value for remaining life-expectancy in the open age group.}

\item{nx.method}{either 2 or 4. 4 is smoother.}

\item{deaths.summed}{logical. is the deaths column given as the total per age in the intercensal period (\code{TRUE}). By default we assume \code{FALSE}, i.e. that the average annual was given.}

\item{mig.summed}{logical. Is the (optional) net migration column \code{mig} given as the total per age in the intercensal period (\code{TRUE}). By default we assume \code{FALSE}, i.e. that the average annual was given.}

\item{delta}{logical. Do we perform the so-called delta adjustment?}

\item{exact.ages.ggb}{optional vector of ages used to estimate GGB coverage (if \code{delta} is \code{TRUE})}

\item{lm.method}{character, one of:\itemize{
\item{\code{"oldschool"}} default sd ratio operation of still unknown origin
\item{\code{"lm"} or \code{"ols"}} for a simple linear model
\item{\code{"tls"}, \code{"orthogonal"}, or \code{"deming"}} for total least squares
\item{\code{"tukey"}, \code{"resistant"}, or "\code{"median"}} for Tukey's resistant line method
}}

\item{opt.method}{what kind of residual do we minimize? choices \code{"RMS"},\code{"logRMS"}, \code{"ORSS"}, \code{"logORSS"} (experimental)}
}
\value{
a \code{data.frame} with columns for the coverage coefficient, and the min and max of the age range on which it is based.
}
\description{
Given two censuses and an average annual number of deaths in each age class between censuses, we can use stable population assumptions to estimate the degree of underregistration of deaths. The method estimates age-specific degrees of coverage. The age pattern of these is assumed to be noisy, so we take the arithmetic mean over some range of ages. One may either specify a particular age-range, or let the age range be determined automatically. If the age-range is found automatically, this is done using the method developed for the generalized growth-balance method. Part of this method relies on a prior value for remaining life expectancy in the open age group. By default, this is estimated using a standard reference to the Coale-Demeny West model life table, although the user may also supply a value. Called by \code{seg()}. Users probably do not need to use this function directly.
}
\details{
Census dates can be given in a variety of ways: 1) using Date classes, and column names \code{$date1} and \code{$date2} (or an unambiguous character string of the date, like, \code{"1981-05-13"}) or 2) by giving column names \code{"day1","month1","year1","day2","month2","year2"} containing integers. If only \code{year1} and \code{year2} are given, then we assume January 1 dates. If year and month are given, then we assume dates on the first of the month.
}
