% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh.R
\name{seggetAgesFit}
\alias{seggetAgesFit}
\title{select age range for SEG method automatically.}
\usage{
seggetAgesFit(codi, minA = 15, maxA = 75, minAges = 8)
}
\arguments{
\item{codi}{\code{data.frame} as produced by \code{ggbMakeColumns()}}

\item{minA}{the lowest age to be included in search}

\item{maxA}{the highest age to be included in search (the lower bound thereof)}

\item{minAges}{the minimum number of adjacent ages to be used in estimating}
}
\value{
integer vector of ages that minimize the RMSE.
}
\description{
We try all possible age ranges given the constraints provided. Whichever age range has the lowest RMSE with respect to the corresponding coverage estimate is returned.
}
