% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh.R
\name{seggetRMS}
\alias{seggetRMS}
\title{calculates RMSE for SEG and a specfied age vector.}
\usage{
seggetRMS(codi, agesFit)
}
\arguments{
\item{codi}{a chunk of data (single sex, year, region, etc) with all columns required by \code{ggb()}}
}
\value{
numeric. RMSE
}
\description{
return the RMSE of `Cx` with respect to the coverage estimate for a given range of ages.
}
