% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh.R
\name{segMakeColumns}
\alias{segMakeColumns}
\title{make the Bennett-Horiuchi quasi life table columns required by the estimation method}
\usage{
segMakeColumns(
  codi,
  minA = 15,
  maxA = 75,
  eOpen = NULL,
  nx.method = 2,
  deaths.summed = FALSE,
  mig.summed = deaths.summed,
  delta = FALSE,
  exact.ages.ggb = NULL,
  lm.method = "oldschool",
  opt.method = "r2",
  minAges.ggb = 8
)
}
\arguments{
\item{codi}{\code{data.frame} with columns, \code{$pop1}, \code{$pop2}, \code{$deaths}, \code{$date1}, \code{$date2}, \code{$sex}, \code{$age}, and \code{$id} (to indicate regions, periods, sexes).}

\item{minA}{the lowest age to be included in search}

\item{maxA}{the highest age to be included in search (the lower bound thereof)}

\item{eOpen}{optional. A user-specified value for remaining life-expectancy in the open age group.}

\item{nx.method}{either 2 or 4. 4 is smoother.}

\item{deaths.summed}{logical. is the deaths column given as the total per age in the intercensal period (\code{TRUE}). By default we assume \code{FALSE}, i.e. that the average annual was given.}

\item{mig.summed}{logical. Is the (optional) net migration column \code{mig} given as the total per age in the intercensal period (\code{TRUE}). By default we assume \code{FALSE}, i.e. that the average annual was given.}

\item{delta}{logical. Do we perform the so-called delta adjustment?}

\item{exact.ages.ggb}{optional vector of ages used to estimate GGB coverage (if \code{delta} is \code{TRUE})}

\item{lm.method}{character, one of:\itemize{
\item{\code{"oldschool"}} default sd ratio operation of still unknown origin
\item{\code{"lm"} or \code{"ols"}} for a simple linear model
\item{\code{"tls"}, \code{"orthogonal"}, or \code{"deming"}} for total least squares
\item{\code{"tukey"}, \code{"resistant"}, or "\code{"median"}} for Tukey's resistant line method
}}

\item{opt.method}{what kind of residual do we minimize? choices \code{"RMS"},\code{"logRMS"}, \code{"ORSS"}, \code{"logORSS"} (experimental)}

\item{minAges.ggb}{the minimum number of adjacent ages to be used in estimating GGB coverage.}
}
\value{
codi, with many columns added, most importantly \code{$Cx}.
}
\description{
Called by \code{segCoverageFromYear()}. This simply modulates some code that would otherwise be repeated. Users probably don't need to call this function directly.
}
