% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abridge.R
\name{single2abr}
\alias{single2abr}
\title{single ages to standard abridged ages}
\usage{
single2abr(x)
}
\arguments{
\item{x}{vector of single ages (lower bound) a.k.a. completed age.}
}
\value{
vector of the same length indicating which abridged age group each single age belongs to (lower bound)
}
\description{
convert ages of the form 0,1,2,3,4,5,... into 0,1,1,1,1,5,...
}
