% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yint}
\alias{yint}
\title{get interval as fraction of full years}
\usage{
yint(Day1, Month1, Year1, Day2, Month2, Year2, reproduce.matlab = FALSE,
  detect.mid.year = TRUE, detect.start.end = TRUE)
}
\arguments{
\item{Day1}{Day of first date}

\item{Month1}{Month of first date}

\item{Year1}{Year of first date}

\item{reproduce.matlab}{logical. default \code{FALSE}. Assume 365 days in all years?}

\item{detect.mid.year}{logical. default \code{TRUE}. Should June 30 and July 1 be considered .5?}

\item{detect.start.end}{logical. default \code{TRUE}. Should Jan 1 always be 0 and Dec 31 always be 1?}

\item{Day1}{Day of second date}

\item{Month1}{Month of second date}

\item{Year1}{Year of second date}
}
\value{
decimal value of year fraction (can be greater than 1)
}
\description{
Either assume 365 days in the year, or get the precise duration.
}

