% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yint2}
\alias{yint2}
\title{get the time interval without having to specify so many args}
\usage{
yint2(X)
}
\arguments{
\item{X}{\code{data.frame} with at least columns \code{$date1} and \code{$date2}, or \code{$year1} and \code{$year2}.}
}
\value{
an decimal year value of the time between two dates.
}
\description{
We accept dates, and fake them otherwise. Dates must be unique. Iterate over data if necessary for multiple intervals.
}

