% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZA}
\alias{ZA}
\title{Example data for South African Males 1997-2007}
\format{
A data frame with 17 rows and 8 variables:
\describe{
  \item{id}{integer a column of 1s}
  \item{sex}{character ``m'' for male}
  \item{pop1}{integer the census population count by age on 2001-10-10}
  \item{pop2}{integer the census population count by age on 2007-02-15}
  \item{deaths}{integer total deaths by age in the intercensal period}
  \item{mig}{integer total net migration by age in the intercensal period}
  \item{age}{integer lower age bound for each age group}
  \item{date1}{Date class, 2001-10-10}
  \item{date2}{Date class, 2007-02-15}
}
}
\source{

}
\usage{
ZA
}
\description{
A dataset containing 18 rows and 8 variables. This includes male population counts from censuses in 2001 and 2007, total deaths between censuses, and total net migration between censuses, all by age.
}
\keyword{datasets}
