% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggb.R
\name{adjustages}
\alias{adjustages}
\title{adjust the range of ages used by \code{ggbChooseAges()}}
\usage{
adjustages(a, age, agesfit)
}
\arguments{
\item{a}{an age specified by the user, as returned by \code{guessage()}}

\item{age}{ages present in dataset}

\item{agesfit}{the former age range used for calculating the coverage coefficient}
}
\value{
the adjusted set of ages used for calculating the coverage coefficient
}
\description{
a utility function called by \code{ggbChooseAges()}. After clicking a point, this function readjusts the age range
}
