% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bh.R
\name{eOpenCD}
\alias{eOpenCD}
\title{estimate remaining life expectancy in the open age group}
\usage{
eOpenCD(codiaugmented)
}
\arguments{
\item{codiaugmented}{the standard codi object being passed through the chain, but having been pre-processed in the course of \code{bh1MakeColumns()} or \code{bh2MakeColumns()}}
}
\value{
numeric an estimate of remaining life expectancy in the open age group
}
\description{
This calculation is based on an indirect method to reference the Coale-Demeny West model lifetable. First one makes a psuedo lifetable deaths column using some stable pop assumptions (different in BH1 vs BH2). Then take the ratio of the sum of ages 10-39 to 40-59. These ratios have been worked out for each model lifetable level, so we can pick the level based on the ratio we produce from the data. From there, we just pick out the remaining life expectancy that corresponds to the top age in our data, which for now hopefully isn't higher than 95. The model lifetables don't go higher than 95 for now, but that's well beyond the range for this method family. If your data go beyong 85 or so, then just group down to 85, say, and estimate using that instead of keeping a high open age. Called by \code{bh1MakeColumns()} and \code{bh2MakeColumns()}, and not intended for direct user interface, because you need to produce the \code{$deathsLT} column. You can skip calling this function by specifying \code{eOpen} in the top call to \code{bh1()} or \code{bh2()}.
}
\details{

}

