% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggb.R
\name{ggbcoverageFromAges}
\alias{ggbcoverageFromAges}
\title{given a set of ages, what is the implied death registration coverage?}
\usage{
ggbcoverageFromAges(codi, agesfit)
}
\arguments{
\item{codi}{a chunk of data (single sex, year, region, etc) with all columns required by \code{ggb()}}

\item{agesfit}{an integer vector of ages, either returned from \code{ggbgetAgesFit} or user-supplied.}
}
\value{
numeric. the estimated level of coverage.
}
\description{
For a single year/sex/region of data (formatted as required by \code{ggb()}), what is the registration coverage implied by a given age range? Called by \code{ggbcoverageFromYear()} and \code{plot.ggb()}.
}

