% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggb.R
\name{ggbcoverageFromYear}
\alias{ggbcoverageFromYear}
\title{estimate death registration coverage for a single year/sex/region using the GGB method}
\usage{
ggbcoverageFromYear(
  codi,
  exact.ages = NULL,
  minA = 15,
  maxA = 75,
  minAges = 8,
  deaths.summed = FALSE,
  mig.summed = deaths.summed,
  lm.method = "oldschool",
  opt.method = "RMS",
  nx.method = 2
)
}
\arguments{
\item{codi}{a chunk of data from a single \code{id}}

\item{exact.ages}{optional. A user-specified vector of exact ages to use for coverage estimation}

\item{minA}{the lowest age to be included in search}

\item{maxA}{the highest age to be included in search (the lower bound thereof)}

\item{minAges}{the minimum number of adjacent ages to be used in estimating}

\item{deaths.summed}{logical. is the deaths column given as the total per age in the intercensal period (\code{TRUE}). By default we assume \code{FALSE}, i.e. that the average annual was given.}

\item{mig.summed}{logical. Is the (optional) net migration column \code{mig} given as the total per age in the intercensal period (\code{TRUE}). By default we assume \code{FALSE}, i.e. that the average annual was given.}

\item{lm.method}{character, one of:\itemize{
\item{\code{"oldschool"}} default sd ratio operation of still unknown origin
\item{\code{"lm"} or \code{"ols"}} for a simple linear model
\item{\code{"tls"}, \code{"orthogonal"}, or \code{"deming"}} for total least squares
\item{\code{"tukey"}, \code{"resistant"}, or "\code{"median"}} for Tukey's resistant line method
}}

\item{opt.method}{what kind of residual do we minimize? choices \code{"RMS"},\code{"logRMS"}, \code{"ORSS"}, \code{"logORSS"} (experimental)}

\item{nx.method}{either 2 or 4. 4 is smoother.}
}
\description{
Given two censuses and an average annual number of deaths in each age class between censuses, we can use stable population assumptions to estimate the degree of underregistration of deaths. The method is based on finding a best-fitting linear relationship between two modeled parameters (right term and left term), but the fit, and resulting coverage estimate, depend on exactly which age range is taken. This function either finds a nice age range for you automatically, or you can specify an exact vector of ages. Called by \code{ggb()}. Users probably don't need to call this directly. Just use \code{ggb()} instead.

Census dates can be given in a variety of ways: 1) using Date classes, and column names \code{$date1} and \code{$date2} (or an unambiguous character string of the date, like, \code{"1981-05-13"}) or 2) by giving column names \code{"day1","month1","year1","day2","month2","year2"} containing integers. If only \code{year1} and \code{year2} are given, then we assume January 1 dates. If year and month are given, then we assume dates on the first of the month.
}
