% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggb.R
\name{ggbgetRMS}
\alias{ggbgetRMS}
\title{calculate the root means square of the error to help find optimal age range}
\usage{
ggbgetRMS(agesi, codi, lm.method = "oldschool")
}
\arguments{
\item{agesi}{the set of ages used for this iteration}

\item{codi}{\code{data.frame} with columns, \code{$pop1}, \code{$pop2}, \code{$deaths}, \code{$date1}, \code{$date2}, and \code{$age}.}

\item{lm.method}{character, one of:\itemize{
\item{\code{"oldschool"}} default sd ratio operation of still unknown origin
\item{\code{"lm"} or \code{"ols"}} for a simple linear model
\item{\code{"tls"}, \code{"orthogonal"}, or \code{"deming"}} for total least squares
\item{\code{"tukey"}, \code{"resistant"}, or "\code{"median"}} for Tukey's resistant line method
}}
}
\value{
the RMSE
}
\description{
Called by \code{ggbgetAgesFit()} whenever the user does not want to manually determine the age range used to determine registration coverage. Probably no need to be called by top-level users. If a user would rather determine the optimal age range some other way, then look to \code{ggbcoverageFromYear()} where \code{ggbgetRMS} is called and add another condition or make it call something else.
}
