% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggb.R
\name{ggbgetRMS}
\alias{ggbgetRMS}
\title{calcuate the root means square of the error to help find optimal age range}
\usage{
ggbgetRMS(agesi, codi)
}
\arguments{
\item{agesi}{the set of ages used for this iteration}

\item{codi}{\code{data.frame} with columns, \code{$pop1}, \code{$pop2}, \code{$deaths}, \code{$date1}, \code{$date2}, and \code{$age}.}
}
\value{
the RMSE
}
\description{
Called by \code{ggbgetAgesFit()} whenever the user doesn't want to manually determine the age range used to determine registration coverage. Probably no need to be called by top-level users. If a user would rather determine the optimal age range some other way, then look to \code{ggbcoverageFromYear()} where \code{ggbgetRMS} is called and add another condition or make it call something else.
}

