% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggb.R
\name{ggbMakeColumns}
\alias{ggbMakeColumns}
\title{make the growth-adjusted quasi lifetable columns required by GGB method}
\usage{
ggbMakeColumns(codi, minA = 15, maxA = 75, deaths.summed = FALSE)
}
\arguments{
\item{codi}{a chunk of data (single sex, year, region, etc) with all columns required by \code{ggb()}}

\item{minA}{the minimum of the age range searched. Default 15}

\item{maxA}{the maximum of the age range searched. Default 75}

\item{deaths.summed}{logical. is the deaths column given as the total per age in the intercensal period (\code{TRUE}). By default we assume \code{FALSE}, i.e. that the average annual was given.}
}
\value{
codi, with many columns added, most importantly \code{$rightterm}, \code{$leftterm}, and \code{$exclude}.
}
\description{
Called by \code{ggbChooseAges()} and \code{ggbcoverageFromYear()}. This simply modulates some code that would otherwise be repeated. Users probably don't need to call this function directly.
}
