% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggb.R
\name{ggbMakeColumns}
\alias{ggbMakeColumns}
\title{make the growth-adjusted quasi life table columns required by GGB method}
\usage{
ggbMakeColumns(
  codi,
  minA = 15,
  maxA = 75,
  nx.method = 2,
  deaths.summed = FALSE,
  mig.summed = deaths.summed
)
}
\arguments{
\item{codi}{a chunk of data (single sex, year, region, etc) with all columns required by \code{ggb()}}

\item{minA}{the lowest age to be included in search}

\item{maxA}{the highest age to be included in search (the lower bound thereof)}

\item{nx.method}{either 2 or 4. 4 is smoother.}

\item{deaths.summed}{logical. is the deaths column given as the total per age in the intercensal period (\code{TRUE}). By default we assume \code{FALSE}, i.e. that the average annual was given.}

\item{mig.summed}{logical. Is the (optional) net migration column \code{mig} given as the total per age in the intercensal period (\code{TRUE}). By default we assume \code{FALSE}, i.e. that the average annual was given.}
}
\value{
codi, with many columns added, most importantly \code{rightterm}, \code{leftterm}, and \code{keep}.
}
\description{
Called by \code{ggbChooseAges()} and \code{ggbcoverageFromYear()}. This simply modulates some code that would otherwise be repeated. Users probably don't need to call this function directly.
}
