% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{headerPrep}
\alias{headerPrep}
\title{a utility function to prep the header}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{cod}{integer an id number for each state}
  \item{pop1}{integer the census population count in 1991}
  \item{pop2}{integer the census population count in 2000}
  \item{deaths}{numeric average deaths between censuses}
  \item{year1}{integer 1991}
  \item{year2}{integer 2000}
  \item{age}{integer lower age bound for each age group}
}}
\source{
DATASUS \url{http://www2.datasus.gov.br/DATASUS/index.php?area=0205}
Example data for Brasil males by federal states, years 1980 to 1991

A dataset containing 486 rows and 7 variables: Population counts for 1980 and 1991 in abridged ages 0, 1, 5, ... 75, with an open age of 80. Deaths are given as the average death count per age group over the intercensal period. In total there are 53 states in this dataset.

DATASUS \url{http://www2.datasus.gov.br/DATASUS/index.php?area=0205}
}
\usage{
headerPrep(X)
}
\arguments{
\item{X}{}
}
\value{
a list of codi chunks (by intercensal period, region, etc), with standardized names, dates, etc.
}
\description{
This is an internal utility function, to save on redundant lines of code. Not so useful for hand-processing.
}
\details{
Example data for Brasil females by federal states, years 1991 to 2000

A dataset containing 486 rows and 7 variables: Population counts for 1991 and 2000 in abridged ages 0, 1, 5, ... 75, with an open age of 80. Deaths are given as the average death count per age group over the intercensal period. In total there are 53 states in this dataset.
}
