% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh.R
\name{seggetAgesFit}
\alias{seggetAgesFit}
\title{select age range for SEG method automatically.}
\usage{
seggetAgesFit(
  codi,
  minA = 15,
  maxA = 75,
  minAges = 8,
  deaths.summed = FALSE,
  nx.method = nx.method
)
}
\arguments{
\item{codi}{a chunk of data (single sex, year, region, etc) with all columns required by \code{ggb()}}

\item{minA}{the lowest age to be included in search}

\item{maxA}{the highest age to be included in search (the lower bound thereof)}

\item{minAges}{the minimum number of adjacent ages to be used in estimating}

\item{deaths.summed}{logical. is the deaths column given as the total per age in the intercensal period (\code{TRUE}). By default we assume \code{FALSE}, i.e. that the average annual was given.}

\item{nx.method}{either 2 or 4. 4 is smoother.}
}
\value{
integer vector of ages that minimize the RMSE.
}
\description{
We try all possible age ranges given the constraints provided. Whichever age range has the lowest RMSE with respect to the corresponding coverage estimate is returned.
}
