% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh.R
\name{seggetRMS}
\alias{seggetRMS}
\title{calculates RMSE for SEG and a specfied age vector.}
\usage{
seggetRMS(codi, agesFit)
}
\arguments{
\item{codi}{\code{data.frame} as returbed by \code{ggbMakeColumns()}}

\item{agesFit}{ineger vector of age trim for SEG method.}
}
\value{
numeric. RMSE
}
\description{
return the RMSE of \code{Cx} with respect to the coverage estimate for a given range of ages.
}
