% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh.R
\name{segMakeColumns}
\alias{segMakeColumns}
\title{make the Bennett-Horiuchi quasi life table columns required by the estimation method}
\usage{
segMakeColumns(codi, minA = 15, maxA = 75, eOpen = NULL, deaths.summed = FALSE)
}
\arguments{
\item{codi}{a chunk of data (single sex, year, region, etc) with all columns required by \code{seg()}}

\item{minA}{the minimum of the age range searched. Default 15}

\item{maxA}{the maximum of the age range searched. Default 75}

\item{eOpen}{optional. A value for remaining life expectancy in the open age group.}

\item{deaths.summed}{logical. is the deaths column given as the total per age in the intercensal period (\code{TRUE}). By default we assume \code{FALSE}, i.e. that the average annual was given.}
}
\value{
codi, with many columns added, most importantly \code{$Cx}.
}
\description{
Called by \code{segCoverageFromYear()}. This simply modulates some code that would otherwise be repeated. Users probably don't need to call this function directly.
}
