% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggb.R
\name{slopeint}
\alias{slopeint}
\title{get the slope the slope and intercept implied by a set of ages}
\usage{
slopeint(codi, agesfit, deaths.summed = FALSE)
}
\arguments{
\item{codi}{\code{data.frame} as produced by \code{ggbMakeColumns()}}

\item{agesfit}{a set of ages to estimate coverage from}

\item{deaths.summed}{logical. is the deaths column given as the total per age in the intercensal period (\code{TRUE}). By default we assume \code{FALSE}, i.e. that the average annual was given.}
}
\value{
a pairlist with elements \code{$a} for the intercept and \code{$b} for the slope
}
\description{
Called by \code{ggbFittedFromAges()} and \code{ggbChooseAges()}
}

