% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggb.R
\name{slopeint}
\alias{slopeint}
\title{get the slope the slope and intercept implied by a set of ages}
\usage{
slopeint(codi, agesfit, lm.method = "oldschool")
}
\arguments{
\item{codi}{\code{data.frame} as produced by \code{ggbMakeColumns()}}

\item{agesfit}{a set of continuous ages to estimate coverage from}

\item{lm.method}{character, one of:\itemize{
  \item{\code{"oldschool"}} default sd ratio operation of still unknown origin
  \item{\code{"lm"} or \code{"ols"}} for a simple linear model
  \item{\code{"tls"}, \code{"orthogonal"}, or \code{"deming"}} for total least squares
  \item{\code{"tukey"}, \code{"resistant"}, or "\code{"median"}} for Tukey's resistant line method
}}
}
\value{
a pairlist with elements \code{$a} for the intercept and \code{$b} for the slope
}
\description{
Called by \code{ggbFittedFromAges()} and \code{ggbChooseAges()}
}
