% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggb.R
\name{slopeint}
\alias{slopeint}
\title{get the slope the slope and intercept implied by a set of ages}
\usage{
slopeint(codi, agesfit)
}
\arguments{
\item{codi}{\code{data.frame} as produced by \code{ggbMakeColumns()}}

\item{agesfit}{a set of ages to estimate coverage from}
}
\value{
a pairlist with elements \code{$a} for the intercept and \code{$b} for the slope
}
\description{
Called by \code{ggbFittedFromAges()} and \code{plot.ggb()}
}

